Minor bug fix to a data type that was incorrectly type cast to int.

+++ LightRail/StationTimes.m
@@ -23,7 +23,7 @@
 //}
 
 
-- (id) initWithStationID:(int)stationID
+- (id) initWithStationID:(long)stationID
 {
     self = [super init];
